build(docker): changed Dockerfile to build from root requirements
This change moves the requirements build to use the requirements.txt
that uses setup.py (setuptools). This should bring the requirments a
developer might add to the repo closer to the deployment process
and reduce any impedance between how the app is deployed versus
how the app is developed.

Any decision to modify the `requirements.txt` in `build/` is left to
a different PR around the infrastructure / deployment process.

QA:
[-] execute a docker build, and confirm no errors
```
docker build -t dsiprouter .
```
[-] execute `deploy.sh docker dev` and confirm no errors
```
deploy.sh docker dev
```
[-] visit site at http://localhost
[-] find a song and serenade yourself on how cool coding is